requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pydantic==1.10.7
pyTelegramBotAPI==4.11.0
python-dotenv==1.0.0
requests==2.30.0
telebot==0.0.5
typing_extensions==4.5.0
urllib3==2.0.2
